Subvert go-vet's printf concern with a non-variadic function
go-vet is complaining because it can recognize that we are calling
`.Get` and ending that call chain with a call to a Printf-like function.

One way to subvert this is to actually call a non-variadic function
before calling Printf.

This error only shows up because we're passing a single non-constant
string to the Printf-like call.